modifié :         .vscode/compile.sh 	modifié :         .vscode/tasks.json 	modifié :         README.md 	nouveau fichier : Read_Projet.html 	nouveau fichier : Zterm.code-workspace 	renommé :         buildevent.zig -> buildTcursed.zig 	nouveau fichier : buildcursed.zig 	supprimé :        deps/curse/clear.zig 	renommé :         deps/curse/event.zig -> deps/curse/cursed.zig 	supprimé :        deps/curse/cursor.zig 	nouveau fichier : deps/curse/dds.zig 	supprimé :        deps/curse/label.zig 	supprimé :        deps/curse/mouse.zig 	supprimé :        deps/curse/style.zig 	supprimé :        deps/curse/term.zig 	modifié :         src_c/Term.cpp 	.vscode/ConvCss.sh 	.vscode/ConvGlade.sh 	.vscode/clearcache.sh 	.vscode/compilelib.sh

+++ .vscode/compile.sh
@@ -17,6 +17,7 @@ projet_lib=$4
 folder_src=$5
 
 
+
 folder_len=${#folder_src}
 
 lib_len=${#projet_lib}
@@ -42,7 +43,17 @@ folder_bin=$lib_projet"zig-out/bin/"$projet_bin
 
 folder_cache=$lib_projet"zig-cache"
 
-echo -en $lib_projet\\n
+folder_cachetest=$lib_projet$folder_src"/zig-cache"
+
+#echo -en $folder_bin\\n
+#echo -en $lib_projet\\n
+#echo -en $projet_src\\n
+#echo -en $folder_src\\n
+#echo -en $projet_typ\\n
+#echo -en $projet_bin\\n
+#echo -en $folder_cachetest\\n
+
+
 #-------------------------------------------------------------------
 # clean
 #-------------------------------------------------------------------
@@ -53,6 +64,11 @@ if test -f $projet_bin ; then
 	rm -r $projet_bin
 fi
 
+if test -d $folder_cachetest ; then
+	rm -r $folder_cachetest
+fi
+
+
 #-------------------------------------------------------------------
 # compile			 Security		Optimizations
 # debug						Yes					No
@@ -62,9 +78,15 @@ fi
 #-------------------------------------------------------------------
 
 if [ "$mode" == "DEBUG" ] ; then
-	( set -x ; \
-				zig build --build-file $lib_projet"build"$projet_src ;\
-	)
+	if [ "$projet_typ" == "ALL" ] ; then
+		( set -x ; \
+					zig build --build-file $lib_projet"build"$projet_src ;\
+		)
+	else
+		(set -x ; \
+					zig test $lib_projet$folder_src"/"$projet_src ;\
+		)
+	fi
 fi
 
 if [ "$mode" == "PROD" ] ; then
@@ -91,6 +113,9 @@ fi
 		ls -lrtsh $folder_bin | cut -d " " -f6
 
 		mv $folder_bin $lib_projet
+		if test -d $folder_cache ; then
 		rm -r $folder_cache
+		fi
 	fi
+
 exit
